+++ sdk/keyvault/keyvault-keys/src/index.ts
@@ -60,7 +60,8 @@ import {
   CryptographyClientOptions,
   LATEST_API_VERSION,
   CreateOctKeyOptions,
-  GetRandomBytesOptions
+  GetRandomBytesOptions,
+  RandomBytes
 } from "./keysModels";
 
 import { CryptographyClient } from "./cryptographyClient";
@@ -167,6 +168,7 @@ export {
   PollOperationState,
   PollerLike,
   PurgeDeletedKeyOptions,
+  RandomBytes,
   RestoreKeyBackupOptions,
   SignOptions,
   SignResult,
@@ -663,10 +665,10 @@ export class KeyClient {
    * @param count - The number of bytes to generate between 1 and 128 inclusive.
    * @param options - The optional parameters.
    */
-  public getRandomBytes(count: number, options: GetRandomBytesOptions = {}): Promise<Uint8Array> {
+  public getRandomBytes(count: number, options: GetRandomBytesOptions = {}): Promise<RandomBytes> {
     return withTrace("getRandomBytes", options, async (updatedOptions) => {
       const response = await this.client.getRandomBytes(this.vaultUrl, count, updatedOptions);
-      return response.value!;
+      return { value: response.value! };
     });
   }